services/ds-workbench/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "datamind-ds-workbench"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "autogluon.tabular>=1.2.0",
    "prophet>=1.1.6",
    "neuralforecast>=2.0.0",
    "dowhy>=0.11.1",
    "econml>=0.15.1",
    "bentoml>=1.3.0",
    "mlflow>=2.18.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "httpx>=0.28.0",
    "langfuse>=2.50.0",
    "structlog>=24.4.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "prometheus-client>=0.21.0",
    "litellm>=1.55.0",
    "tenacity>=9.0.0",
    "shap>=0.46.0",
    "minio>=7.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-benchmark>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ds_workbench"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "S"]
ignore = ["S101", "S301", "S603", "S607", "S311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
